Ensure each log entry writes to LevelUP in series
Due to the way the previous log-to-LevelUP pipeline chained Transform
streams, it was possible for a log segment like:

    ... Save Form F ... Publish Form F as Project P ...

To produce an error because Form F wasn't written to LevelUP by the
time the pipeline tried to fetch Form F from LevelUP to create records
for Project P.  api.commonform.org ran into this issue with a number of
log segments where saving a form was immediately followed by publication
of the form.

This commit simplifies the log-to-LevelUP pipeline, most importantly
by ensuring that transformation from log entry to LevelUP batch
and writing of the batch happen within a single Transform stream.
That Transform only calls back for an entry chunk once it has written
the corresponding LevelUP batch.